Revert "Change default of upgrade to FALSE"
This reverts commit e392710e5f20123b1310eff0bc18a901d0a29816.

The upgrade logic does not currently handle the version requirements of
a package. So if you have v1.0.0 of package abc try to install package
xlz with `upgrade = FALSE` and xyz requires version v2.0.0 the install
will fail.

It would be possible to modify code to change this, but it would require
changes in at least `package_deps()`, `local_package_deps()`,
`dev_package_deps()`, `compare_versions()` and `update()` and I do not
think it worth the investment at this time.

Fixes #183