Revert "Change default of upgrade to FALSE"
This reverts commit 42045fcc33ee2ad800a95568983c2beae1eab296.

The upgrade logic does not currently handle the version requirements of
a package. So if you have v1.0.0 of package abc try to install package
xlz with `upgrade = FALSE` and xyz requires version v2.0.0 the install
will fail.

It would be possible to modify code to change this, but it would require
changes in at least `package_deps()`, `local_package_deps()`,
`dev_package_deps()`, `compare_versions()` and `update()` and I do not
think it worth the investment at this time.

Fixes #183