Skip fetching submodules which exist in .Rbuildignore
This code is more complicated than it would otherwise need to be because
the way that `R CMD build` excludes files means that having `^foo$` in
your .Rbuildignore will ignore _all_ files in the foo/ directory.

So to determine if a given path should be ignored we need to match both
the path and it's directories against the patterns.

Fixes https://github.com/r-lib/remotes/issues/201