Fix retrieval of package name on Windows
`writeLines` opens files in text mode, which means on windows newlines are
'\r\n', however read.dcf expects '\n' newlines, so you get whitespace
errors due to the extra blank lines. `writeChar` opens files in binary
mode, so no newline conversion is done, and in this case the whole
description is in one vector, so there is no need for writeLines.

Fixes #248