Custom headers in HTTP, use headers for basic auth
BitBucket does not support Basic Auth via URL encoding,
apparently. So we use HTTP headers instead. This is somewhat
tricky, because base R  download.file()) does not allow
that directly. We need to shadow an internal function in utils,
and also make use of a bugs in curl and the internal
download methods, that let you specify multiple HTTP headers
via the User-Agent option.

This will also fix #255.