Install child dependencies before parents
Previously the parent dependencies were first in the list of packages
passed down to `install.packages()`, which caused them to be installed
in the wrong order when there was a mix of binary and source packages
available. Putting the child dependencies first ensures the order will
be correct even in this situation.

Fixes #296